backend/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
